Moved Dynamic components, input factory service and ui config model to separate module
