+++ client/app/shared/UIConfig.model.ts
