+++ client/app/shared/group/group.component.scss
