+++ client/app/shared/group/group.component.spec.ts
