+++ client/app/shared/input-factory.service.ts
