+++ client/app/shared/label/label.component.scss
