+++ client/app/shared/label/label.component.spec.ts
