+++ client/app/shared/label/label.component.ts
