+++ client/app/shared/tab/tab.component.ts
