+++ client/app/shared/tag/tag.component.ts
