+++ client/app/step-builder/step-builder/step-builder.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { UIConfig } from '../shared/UIConfig.model';
+import { UIConfig } from '../../shared/UIConfig.model';
 
 @Component({
   selector: 'app-step-builder',
